Move test entities to io.realm.entities package
