+++ test/base/src/androidTest/kotlin/io/realm/test/shared/EncryptionTests.kt
@@ -18,8 +18,8 @@ package io.realm.test.shared
 
 import io.realm.Realm
 import io.realm.RealmConfiguration
+import io.realm.entities.Sample
 import io.realm.test.platform.PlatformUtils
-import test.Sample
 import kotlin.random.Random
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
