+++ test/base/src/androidTest/kotlin/io/realm/test/shared/ImportTests.kt
@@ -18,13 +18,13 @@ package io.realm.test.shared
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmObject
+import io.realm.entities.Sample
+import io.realm.entities.link.Child
+import io.realm.entities.link.Parent
 import io.realm.isManaged
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.TypeDescriptor.classifiers
 import io.realm.test.util.Utils.createRandomString
-import test.Sample
-import test.link.Child
-import test.link.Parent
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
