+++ test/base/src/androidTest/kotlin/io/realm/test/shared/MutableRealmTests.kt
@@ -20,9 +20,9 @@ import io.realm.RealmConfiguration
 import io.realm.objects
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.Utils.createRandomString
-import test.StringPropertyWithPrimaryKey
-import test.link.Child
-import test.link.Parent
+import io.realm.entities.StringPropertyWithPrimaryKey
+import io.realm.entities.link.Child
+import io.realm.entities.link.Parent
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
