+++ test/base/src/androidTest/kotlin/io/realm/test/shared/NullabilityTests.kt
@@ -17,9 +17,9 @@ package io.realm.test.shared
 
 import io.realm.Realm
 import io.realm.RealmConfiguration
+import io.realm.entities.Nullability
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.Utils.createRandomString
-import test.Nullability
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
