+++ test/base/src/androidTest/kotlin/io/realm/test/shared/PrimaryKeyTests.kt
@@ -18,23 +18,23 @@ package io.realm.test.shared
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmObject
+import io.realm.entities.primarykey.NoPrimaryKey
+import io.realm.entities.primarykey.PrimaryKeyByte
+import io.realm.entities.primarykey.PrimaryKeyByteNullable
+import io.realm.entities.primarykey.PrimaryKeyChar
+import io.realm.entities.primarykey.PrimaryKeyCharNullable
+import io.realm.entities.primarykey.PrimaryKeyInt
+import io.realm.entities.primarykey.PrimaryKeyIntNullable
+import io.realm.entities.primarykey.PrimaryKeyLong
+import io.realm.entities.primarykey.PrimaryKeyLongNullable
+import io.realm.entities.primarykey.PrimaryKeyShort
+import io.realm.entities.primarykey.PrimaryKeyShortNullable
+import io.realm.entities.primarykey.PrimaryKeyString
+import io.realm.entities.primarykey.PrimaryKeyStringNullable
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.TypeDescriptor.allPrimaryKeyFieldTypes
 import io.realm.test.util.TypeDescriptor.rType
 import io.realm.test.util.Utils.createRandomString
-import test.primarykey.NoPrimaryKey
-import test.primarykey.PrimaryKeyByte
-import test.primarykey.PrimaryKeyByteNullable
-import test.primarykey.PrimaryKeyChar
-import test.primarykey.PrimaryKeyCharNullable
-import test.primarykey.PrimaryKeyInt
-import test.primarykey.PrimaryKeyIntNullable
-import test.primarykey.PrimaryKeyLong
-import test.primarykey.PrimaryKeyLongNullable
-import test.primarykey.PrimaryKeyShort
-import test.primarykey.PrimaryKeyShortNullable
-import test.primarykey.PrimaryKeyString
-import test.primarykey.PrimaryKeyStringNullable
 import kotlin.reflect.typeOf
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
