+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmListTests.kt
@@ -22,15 +22,15 @@ import io.realm.RealmConfiguration
 import io.realm.RealmList
 import io.realm.RealmObject
 import io.realm.RealmResults
+import io.realm.entities.list.Level1
+import io.realm.entities.list.Level2
+import io.realm.entities.list.Level3
+import io.realm.entities.list.RealmListContainer
 import io.realm.objects
 import io.realm.realmListOf
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.TypeDescriptor
 import io.realm.toRealmList
-import test.list.Level1
-import test.list.Level2
-import test.list.Level3
-import test.list.RealmListContainer
 import kotlin.reflect.KClassifier
 import kotlin.reflect.KMutableProperty1
 import kotlin.test.AfterTest
