+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmResultsTests.kt
@@ -19,10 +19,10 @@ import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmResults
 import io.realm.VersionId
+import io.realm.entities.link.Child
+import io.realm.entities.link.Parent
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.Utils.createRandomString
-import test.link.Child
-import test.link.Parent
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
