+++ test/base/src/androidTest/kotlin/io/realm/test/shared/notifications/RealmNotificationsTests.kt
@@ -19,13 +19,13 @@ package io.realm.test.shared.notifications
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.VersionId
+import io.realm.entities.Sample
 import io.realm.internal.platform.runBlocking
 import io.realm.test.NotificationTests
 import io.realm.test.platform.PlatformUtils
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.collect
-import test.Sample
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Ignore
