+++ test/base/src/androidTest/kotlin/io/realm/test/shared/notifications/RealmObjectNotificationsTests.kt
@@ -18,6 +18,7 @@ package io.realm.test.shared.notifications
 
 import io.realm.Realm
 import io.realm.RealmConfiguration
+import io.realm.entities.Sample
 import io.realm.observe
 import io.realm.test.NotificationTests
 import io.realm.test.platform.PlatformUtils
@@ -27,7 +28,6 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.onCompletion
 import kotlinx.coroutines.runBlocking
-import test.Sample
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Ignore
