+++ test/base/src/androidTest/kotlin/io/realm/test/shared/notifications/SystemNotificationTests.kt
@@ -18,11 +18,11 @@ package io.realm.test.shared.notifications
 
 import io.realm.Realm
 import io.realm.RealmConfiguration
+import io.realm.entities.Sample
 import io.realm.internal.platform.singleThreadDispatcher
 import io.realm.test.platform.PlatformUtils
 import io.realm.test.util.Utils
 import kotlinx.coroutines.runBlocking
-import test.Sample
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
