+++ test/base/src/commonMain/kotlin/io/realm/entities/list/RealmListContainer.kt
@@ -9,13 +9,12 @@
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied.
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-package test.list
+package io.realm.entities.list
 
 import io.realm.RealmList
 import io.realm.RealmObject
